+++ tests/testthat/test_MSnSet.R
@@ -221,18 +221,6 @@ test_that("makeImpuritiesMatrix", {
     expect_equal(a6, m6, check.attributes = FALSE)
 })
 
-test_that("Normalisation and transpose", {
-    data(msnset)
-    msnset1 <- normalise(msnset, "sum")
-    expect_equivalent(rowSums(exprs(msnset1), na.rm = TRUE),
-                      rep(1, nrow(msnset1)))
-    msnset2 <- normalise(msnset, "max")
-    expect_true(all(apply(exprs(msnset2), 1, max, na.rm = TRUE) == 1))
-    msnset3 <- normalise(msnset, "quantiles")
-    msnset4 <- normalise(msnset, "quantiles.robust")
-    msnset5 <- normalise(msnset, "vsn")
-})
-
 
 test_that("Transpose and subset", {
     data(msnset)